--- src/background/DataIngestion/Functions/CoPilotMetricsIngestion/Models/DotComPullRequestRepository.cs
@@ -1,13 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace Microsoft.CopilotDashboard.DataIngestion.Functions.CoPilotMetricsIngestion.Models;
-
-public class DotComPullRequestRepository
-{
-    [JsonPropertyName("name")]
-    public required string Name { get; set; }
-
-    [JsonPropertyName("total_engaged_users")]
-    public required int TotalEngagedUsers { get; set; }
-
-}
