--- src/background/DataIngestion/Functions/CoPilotUsageIngestion/TestData/TestData.json
@@ -1,42 +1,0 @@
-﻿[
-  {
-    "total_suggestions_count": 8312,
-    "total_acceptances_count": 2009,
-    "total_lines_suggested": 35290,
-    "total_lines_accepted": 5038,
-    "total_active_users": 97,
-    "total_chat_acceptances": 157,
-    "total_chat_turns": 190,
-    "total_active_chat_users": 38,
-    "day": "2024-05-14",
-    "breakdown": [
-      {
-        "language": "dockercompose",
-        "editor": "vscode",
-        "suggestions_count": 7,
-        "acceptances_count": 2,
-        "lines_suggested": 133,
-        "lines_accepted": 66,
-        "active_users": 2
-      },
-      {
-        "language": "git-commit",
-        "editor": "vscode",
-        "suggestions_count": 68,
-        "acceptances_count": 3,
-        "lines_suggested": 90,
-        "lines_accepted": 3,
-        "active_users": 4
-      },
-      {
-        "language": "github-actions-workflow",
-        "editor": "vscode",
-        "suggestions_count": 191,
-        "acceptances_count": 51,
-        "lines_suggested": 355,
-        "lines_accepted": 148,
-        "active_users": 4
-      }
-    ]
-  }
-]
